Further improved performance by 22%
This is mainly due to submitting individual tasks for each triangle to the pool instead of pushing groups: they have different sizes, so they should be scheduled individually.

Other work includes allocation reductions, because I like to do these, because they remove a bit of strain from the system in low memory situations.